+++ Source/Core/DotSpatial.Controls/MapLabelLayer.cs
@@ -667,9 +667,13 @@ namespace DotSpatial.Controls
                 {
                     var ls = GetSegment(ls1, symb);
                     if (ls == null) return 0;
+
                     if (symb.LineOrientation == LineOrientation.Parallel)
-                        return ToSingle(-ls.Angle);
-                    return ToSingle(-ls.Angle - 90);
+                    {
+                        return ToSingle(-GetDisplayAngle(ls, 0));
+                    }
+
+                    return ToSingle(-GetDisplayAngle(ls, -Math.PI / 2));
                 }
             }
 
@@ -713,6 +717,28 @@ namespace DotSpatial.Controls
             }
 
             return null;
+        }
+
+        /// <summary>
+        /// Gets the angle that will be used to draw line based labels so they won't be on their head.
+        /// </summary>
+        /// <param name="ls">Linestring, that is used to determine the display angle of the label.</param>
+        /// <param name="add">Angle in radian that needs to be added for the label to be rotated correctly (e.g. perpendicular labels get rotated by -90°)</param>
+        private static double GetDisplayAngle(LineSegment ls, double add)
+        {
+            double deltaX = ls.P1.X - ls.P0.X;
+            double deltaY = ls.P1.Y - ls.P0.Y;
+            double length = Math.Sqrt(deltaX * deltaX + deltaY * deltaY);
+            double angleRAD = Math.Asin(Math.Abs(ls.P1.Y - ls.P0.Y) / length) + add;
+
+            double angle = (angleRAD * 180) / Math.PI;
+
+            if (((ls.P0.X < ls.P1.X) && (ls.P0.Y > ls.P1.Y)) || ((ls.P0.X > ls.P1.X) && (ls.P0.Y < ls.P1.Y)))
+            {
+                angle = 360 - angle;
+            }
+
+            return angle;
         }
 
         /// <summary>